+++ eslint.config.mjs
@@ -3,11 +3,15 @@ import tseslint from "typescript-eslint";
 
 export default tseslint.config(
   {
-    ignores: ["**/dist/**", "**/node_modules/**", "eslint.config.mjs"],
+    ignores: ["**/dist/**", "**/node_modules/**", "coverage/**", "eslint.config.mjs"],
   },
   eslint.configs.recommended,
   ...tseslint.configs.strictTypeChecked,
   ...tseslint.configs.stylisticTypeChecked,
+  {
+    files: ["tests/**/*.mjs"],
+    ...tseslint.configs.disableTypeChecked,
+  },
   {
     files: ["**/*.ts"],
     languageOptions: {